增加电子拍牌数据上传

+++ wintop-carauction-service/src/main/java/com/wintop/ms/carauction/model/TblAuctionLogModel.java
@@ -33,7 +33,7 @@ public class TblAuctionLogModel {
     /**
      * 根据主键查询记录
      */
-    public TblAuctionLog selectByPrimaryKey(Integer id){
+    public TblAuctionLog selectByPrimaryKey(Long id){
         return tblAuctionLogReadDao.selectByPrimaryKey(id);
     }
